[Blazor] Adds a custom packaging sample
This commit adds a sample that demonstrates how the blazor publish output
can be customized to enable custom bootstrapping of Blazor applications.

This commit includes two new packages 'Microsoft.AspNetCore.Components.WebAssembly.Packaging'
and 'Microsoft.AspNetCore.Components.WebAssembly.Packaging.Server'that
work together to enable customizing the publish output for Blazor and the
subsequent customized loading process by the browser.

Microsoft.AspNetCore.Components.WebAssembly.Packaging contains MSBuild
targets to attach to the Blazor publish process and generate a single
multipart bundle file that is consumed by a JS initializer also defined
in this package and that is automatically imported by Blazor WebAssembly
when the application starts.

The JS initializer included in the package, downloads the bundle and creates
an object url for each one of the files included in the bundle that then
provides to the resource loader when asked for them.

Microsoft.AspNetCore.Components.WebAssembly.Packaging.Server contains a
helper method to streamline the serving of the application bundle and its
compressed versions when supported, and its not technically necessary.