Add readonly remote session
This change implements the remote session management against the
adapters. This only enables it for readonly state at the moment, but
ties it together so users can access `HttpContext.Session["key"]` and
get the type they expect. A registration is set up to map keys to types
so they can be serialized/deserialized correctly.