Add instrumentation to adapters
This enables the ability to track how much effect the adapters are having on the execution of endpoints. This will also allow us to identify unnecessary work that is being done (i.e. session is turned on, but not being used), as well as how much reliance an endpoint has on the adapters.